Adds padding customization to breadcrumb items
Allows customization of padding for breadcrumb items
and page items (current page) via the theme.

This change introduces `itemPadding` and `pagePadding` properties
to the `ShadBreadcrumbTheme`, enabling developers to control the
spacing around breadcrumb elements. Padding is only applied if
the breadcrumb item is not a link or a button, as those components
already have their own built-in padding.