fix: popover cannot be closed with esc key
When the popover is opened via a mouse click, it did not gain focus,
preventing it from receiving keyboard events. As a result, the 'ESC' key
shortcut to close it did not work.

This commit fixes the issue by programmatically requesting focus for the
popover as soon as it becomes visible.

- A `FocusNode` is attached to the popover content to make it focusable.
- A `FocusScope` is introduced to trap focus within the popover.